feat: add summoner lookup and codebase separation
