+++ opgg/v2/search_result.py
@@ -11,11 +11,11 @@ class SearchResult(Box):
 
     ##### Properties:
         `region`: The region of the summoner.
-        `summoner_result`: The summoner result.
+        `summoner`: The summoner result.
     """
 
     region: Region
-    summoner_result: Summoner
+    summoner: Summoner
 
     def __str__(self):
-        return f"[{str(self.region):4}] {f'{self.summoner_result.game_name} #{self.summoner_result.tagline}':<25} | Level: {self.summoner_result.level:<4} [Summoner ID: {self.summoner_result.summoner_id}]"
+        return f"[{str(self.region):4}] {f'{self.summoner.game_name} #{self.summoner.tagline}':<25} | Level: {self.summoner.level:<4} [Summoner ID: {self.summoner.summoner_id}]"
